Move notebook template-filling code into nb subpackage
