+++ geoAssembler/nb/templates/geoAssembler.tmpl
